chore(tests): split monolithic JsonApiTest into JsonApi/ subdir
Replace the 8-method root JsonApiTest.php with 4 focused JsonApi/* test
files, mirroring the JsonLd/ and Hal/ subdir layout:

* JsonApi/ErrorTest (ErrorResource rendered in vnd.api+json)
* JsonApi/IdentifierModeTest (4 use_iri_as_id:false tests sharing the
  bootJsonApiKernel helper)
* JsonApi/IriModeTest (default use_iri_as_id:true mode)
* JsonApi/InputDtoTest (POST with input DTOs, preserved attributes and
  required constructor args)

Each file scopes its fixtures to only what its tests use.
